ported BroadcastConsensus2 deleted the old implementation
