+++ Source/VCGeneration/Split.cs
@@ -1153,10 +1153,7 @@ namespace VC
         VCExpr vc = parent.GenerateVCAux(impl, controlFlowVariableExpr, label2absy, checker.TheoremProver.Context);
         Contract.Assert(vc != null);
         
-        if (InstantiationSourceChecker.HasInstantiationSources(impl))
-        {
-          vc = Microsoft.Boogie.VCExprAST.QuantifierInstantiationEngine.Instantiate(impl, exprGen, bet, vc);
-        }
+        vc = QuantifierInstantiationEngine.Instantiate(impl, exprGen, bet, vc);
 
         VCExpr controlFlowFunctionAppl =
           exprGen.ControlFlowFunctionApplication(exprGen.Integer(BigNum.ZERO), exprGen.Integer(BigNum.ZERO));
