+++ Test/civl/inductive-sequentialization/BroadcastConsensus2.bpl.expect
@@ -1,0 +1,2 @@
+
+Boogie program verifier finished with 29 verified, 0 errors
